requirements.txt
bittensor==9.9.0
torch>=2.0
torchvision
kaggle
pillow
timm>=1.0.9